Split up the terminal tests
